Fixed issue with arg parser usign default dates that weren't exactly 1 day apart.

+++ convert_history.py
@@ -32,12 +32,13 @@ from pytz import all_timezones
 from tzlocal import get_localzone
 import os
 
+
 def change_timezone(col, timezone):
     # Must have tz set otherwise will fail
     try:
         convert_col = col.dt.tz_convert(tz=timezone)
     except AttributeError as e:
-        logging.error("The column is not a date time type or there is no time zone from which to converet.")
+        logging.error(f"The column is not a date time type or there is no time zone from which to convert. Error {e}")
         convert_col = col
     return convert_col
 
@@ -46,7 +47,7 @@ def timestamp_to_date(col):
     return pd.to_datetime(col, origin='unix', unit='ms', utc=True, errors="coerce")
 
 
-def convert_history(data_file, timezone):
+def convert_history(data_file, timezone, keep_original):
     logging.debug(f"Converting data file {data_file} from timestamp to local timezone.")
     date_cols = ["sourcetimestamp", "firstactiveat", "changedon"]
     try:
@@ -58,11 +59,11 @@ def convert_history(data_file, timezone):
     df[date_cols] = df[date_cols].replace(to_replace=0, value=np.nan)
     df.update(df[date_cols].apply(pd.to_datetime, origin='unix', unit='ms', utc=True, errors="coerce"))
     df.update(df[date_cols].apply(change_timezone, args=(timezone,)))
-    try:
-        os.rename(data_file, data_file + ".old")
-    except IOError as e:
-        logging.error(f"Tried to rename old file {data_file}.old but failed with error{e}")
-#    new_filename = data_file.replace(".csv", "-converted.csv")
+    if keep_original:
+        try:
+            os.rename(data_file, data_file + ".old")
+        except IOError as e:
+            logging.error(f"Tried to rename old file {data_file}.old but failed with error{e}")
     try:
         df.to_csv(data_file, date_format='%Y-%m-%d %H:%M:%S.%f'[:-3])
         logging.info(f"Converted file written to {data_file}.")
@@ -82,6 +83,8 @@ if __name__ == '__main__':
     parser.add_argument("-tz", "--timezone",
                         dest="timezone",
                         help="Time zone offset in hours minutes HH:MM e.g. 10:00 or -4:00")
+    parser.add_argument("-ko", "--keep_original", dest="keep_original", default=False, action='store_false',
+                        help="Keep the original file with timestamps as .old")
     args = parser.parse_args()
     if args.timezone is None:
         tz = get_localzone()
@@ -91,4 +94,4 @@ if __name__ == '__main__':
         logging.error(f"Timezone {args.timezone} is not valid. Using local timezone {tz}")
     else:
         tz = args.timezone
-    convert_history(args.filename, tz)
+    convert_history(args.filename, tz, args.keep_original)
